First and probably last version of this file
The idea was to write sentiment analysis for Twitter, but it turns out
that people suck and the worst of them gather on Twitter. I needed a
good, active #hashtag to use as the search term, which mostly just
includes Swedish, so I thought I'd go with #svpol. What could possibly
go wrong (tm).

Turns out #svpol is filled to the brim with racists, illiterates and
a generous assortment of other assholes. I just can't bring myself to
work with this dataset.

Also, I realized that I'd have to manually label an absolute shitton
of data to get any kind of useful results. Perhaps I'll resume this
project sometime in the future.