Save tweets in sqlite database
Well, I guess I was wrong - I've resumed working on this file. It
turns out it was too interesting to resist. Now, it will save the
tweet sentences in an sqlite database, and I've added functionality
to EnforsBot to read that database and let the user provide labels
for later sentiment analysis of future incoming tweets.